refactor: restore runtimes-defined features type in controllers & original QServiceManagerFeatures
