+++ server/aws-lsp-codewhisperer/src/language-server/chat/telemetry/chatTelemetryController.ts
@@ -7,7 +7,7 @@ import {
     CombinedConversationEvent,
     InteractWithMessageEvent,
 } from '../../../shared/telemetry/types'
-import { KeysMatching } from '../../types'
+import { Features, KeysMatching } from '../../types'
 import {
     ChatUIEventName,
     InsertToCursorPositionParams,
@@ -22,7 +22,6 @@ import { AcceptedSuggestionEntry, CodeDiffTracker } from '../../inline-completio
 import { TelemetryService } from '../../../shared/telemetry/telemetryService'
 import { getEndPositionForAcceptedSuggestion } from '../../../shared/utils'
 import { CodewhispererLanguage } from '../../../shared/languageDetection'
-import { Features } from '../../../shared/amazonQServiceManager/BaseAmazonQServiceManager'
 
 export const CONVERSATION_ID_METRIC_KEY = 'cwsprChatConversationId'
