+++ server/aws-lsp-codewhisperer/src/language-server/configuration/qConfigurationServer.ts
@@ -52,8 +52,6 @@ export const QConfigurationServerToken =
                 runtime,
                 workspace,
                 sdkInitializator,
-                chat,
-                telemetry,
             })
 
             serverConfigurationProvider = new ServerConfigurationProvider(
@@ -63,10 +61,10 @@ export const QConfigurationServerToken =
             )
 
             /* 
-                    Calling handleDidChangeConfiguration once to ensure we get configuration atleast once at start up
-                    
-                    TODO: TODO: consider refactoring such responsibilities to common service manager config/initialisation server
-                */
+                        Calling handleDidChangeConfiguration once to ensure we get configuration atleast once at start up
+                        
+                        TODO: TODO: consider refactoring such responsibilities to common service manager config/initialisation server
+                    */
             await amazonQServiceManager.handleDidChangeConfiguration()
         })
