+++ server/aws-lsp-codewhisperer/src/language-server/inline-completion/codeWhispererServer.ts
@@ -29,7 +29,10 @@ import { textUtils } from '@aws/lsp-core'
 import { TelemetryService } from '../../shared/telemetry/telemetryService'
 import { AcceptedSuggestionEntry, CodeDiffTracker } from './codeDiffTracker'
 import { AmazonQError, AmazonQServiceInitializationError } from '../../shared/amazonQServiceManager/errors'
-import { AmazonQBaseServiceManager, Features } from '../../shared/amazonQServiceManager/BaseAmazonQServiceManager'
+import {
+    AmazonQBaseServiceManager,
+    QServiceManagerFeatures,
+} from '../../shared/amazonQServiceManager/BaseAmazonQServiceManager'
 import { initBaseTokenServiceManager } from '../../shared/amazonQServiceManager/AmazonQTokenServiceManager'
 import { AmazonQWorkspaceConfig } from '../../shared/amazonQServiceManager/configurationUtils'
 import { initBaseIAMServiceManager } from '../../shared/amazonQServiceManager/AmazonQIAMServiceManager'
@@ -224,7 +227,7 @@ interface AcceptedInlineSuggestionEntry extends AcceptedSuggestionEntry {
 }
 
 export const CodewhispererServerFactory =
-    (serviceManager: (features: Features) => AmazonQBaseServiceManager): Server =>
+    (serviceManager: (features: QServiceManagerFeatures) => AmazonQBaseServiceManager): Server =>
     ({ credentialsProvider, lsp, workspace, telemetry, logging, runtime, sdkInitializator }) => {
         let lastUserModificationTime: number
         let timeSinceLastUserModification: number = 0
@@ -571,9 +574,9 @@ export const CodewhispererServerFactory =
             codePercentageTracker.customizationArn = customizationArn
             logging.debug(`CodePercentageTracker customizationArn updated to ${customizationArn}`)
             /*
-                    The flag enableTelemetryEventsToDestination is set to true temporarily. It's value will be determined through destination
-                    configuration post all events migration to STE. It'll be replaced by qConfig['enableTelemetryEventsToDestination'] === true
-                */
+                            The flag enableTelemetryEventsToDestination is set to true temporarily. It's value will be determined through destination
+                            configuration post all events migration to STE. It'll be replaced by qConfig['enableTelemetryEventsToDestination'] === true
+                        */
             // const enableTelemetryEventsToDestination = true
             // telemetryService.updateEnableTelemetryEventsToDestination(enableTelemetryEventsToDestination)
             telemetryService.updateOptOutPreference(optOutTelemetryPreference)
@@ -588,7 +591,6 @@ export const CodewhispererServerFactory =
                 runtime,
                 sdkInitializator,
                 workspace,
-                telemetry,
             })
 
             const clientParams = safeGet(
@@ -620,10 +622,10 @@ export const CodewhispererServerFactory =
             )
 
             /* 
-                    Calling handleDidChangeConfiguration once to ensure we get configuration atleast once at start up
-                    
-                    TODO: TODO: consider refactoring such responsibilities to common service manager config/initialisation server
-                */
+                            Calling handleDidChangeConfiguration once to ensure we get configuration atleast once at start up
+                            
+                            TODO: TODO: consider refactoring such responsibilities to common service manager config/initialisation server
+                        */
             await amazonQServiceManager.handleDidChangeConfiguration()
             await amazonQServiceManager.addDidChangeConfigurationListener(updateConfiguration)
         }
