+++ server/aws-lsp-codewhisperer/src/language-server/securityScan/codeWhispererSecurityScanServer.ts
@@ -239,14 +239,13 @@ export const SecurityScanServerToken =
                 credentialsProvider,
                 sdkInitializator,
                 workspace,
-                telemetry,
             })
             scanHandler = new SecurityScanHandler(amazonQServiceManager, workspace, logging)
             /* 
-                    Calling handleDidChangeConfiguration once to ensure we get configuration atleast once at start up
-                    
-                    TODO: TODO: consider refactoring such responsibilities to common service manager config/initialisation server
-                */
+                        Calling handleDidChangeConfiguration once to ensure we get configuration atleast once at start up
+                        
+                        TODO: TODO: consider refactoring such responsibilities to common service manager config/initialisation server
+                    */
             await amazonQServiceManager.handleDidChangeConfiguration()
         }
